wrap query fn in runtime type assert
