+++ src/scraper/flow.ts
@@ -3,7 +3,8 @@ import * as Rx from 'rxjs'
 import * as ops from 'rxjs/operators'
 import * as RxCustom from '../util/rxjs/observables'
 import { ScrapeStep } from './scrape-step'
-import { wrapError, ResponseError } from '../util/error'
+import { wrapError } from '../util/rxjs/operators'
+import { ResponseError } from '../util/error'
 // type imports
 import { Tools } from '../tools'
 import { FMap } from '../util/map'
