+++ src/settings/config/assert.runtime.ts
@@ -1,11 +1,7 @@
 import { assertType } from 'typescript-is'
 import { ConfigInit } from './types'
-import { RuntimeTypeError } from '../../util/error'
+import { typescriptIsWrapper } from '../../util/error'
 
-export const assertConfigType = (configInit: any) => {
-  try {
-    assertType<ConfigInit>(configInit)
-  } catch (e) {
-    throw new RuntimeTypeError(e.message)
-  }
-}
+export const assertConfigType = typescriptIsWrapper(configInit =>
+  assertType<ConfigInit>(configInit)
+)
