+++ src/settings/params/assert.runtime.ts
@@ -1,12 +1,7 @@
 import { assertType } from 'typescript-is'
 import { ParamsInit } from './types'
-import { RuntimeTypeError } from '../../util/error'
+import { typescriptIsWrapper } from '../../util/error'
 
-/* eslint-disable-next-line @typescript-eslint/no-unused-vars */
-export const assertParamsType = (paramsInit: any) => {
-  try {
-    assertType<ParamsInit>(paramsInit)
-  } catch (e) {
-    throw new RuntimeTypeError(e.message)
-  }
-}
+export const assertParamsType = typescriptIsWrapper(paramsInit =>
+  assertType<ParamsInit>(paramsInit)
+)
