+++ src/tools/store/index.ts
@@ -1,14 +1,14 @@
 import { Database } from './database'
-import { groupUntilSeparator } from '../../util/array'
 import { Config, FlatConfig } from '../../settings/config/types'
-import { createTables, createStatements, selectOrderedScrapers } from './queries'
+import { createTables, createStatements } from './queries'
+import { querierFactory } from './querier-entrypoint'
 // type imports
 import { Transaction } from 'better-sqlite3'
 import { Settings } from '../../settings'
-import { ScraperName } from '../../settings/config/types'
-import { SelectedRow as OrderedScrapersRow } from './queries/select-ordered-scrapers'
 
 class Store {
+  public static querierFactory = querierFactory
+
   public qs: ReturnType<typeof createStatements>
   public transaction: Transaction
   private config: Config
@@ -27,51 +27,6 @@ class Store {
     // prepare statements
     this.qs = createStatements(this.flatConfig, this.database)
   }
-
-  /**
-   * external function for grabbing data back out of the scraper
-   * the first time the produced function is called, it will create the database & sql query
-   * this statefull nonsense is necessary so we can give the user `query` before awaiting on folder creation
-   */
-  public static getQuerier = (settings: Settings): QueryFn => {
-    const { flatConfig, paramsInit } = settings
-    let firstCall = true
-    let database: Database
-
-    const prepare: QueryFn['prepare'] = ({ scrapers, groupBy }) => {
-      if (firstCall) {
-        // this stateful stuff is necessary so we can give this to the user before creating folders
-        database = new Database(paramsInit.folder)
-      }
-      if (!scrapers.some(s => flatConfig.has(s))) return () => []
-
-      const scrapersInConfig = scrapers.concat(groupBy || []).filter(s => flatConfig.has(s))
-
-      const preparedStatment = selectOrderedScrapers(flatConfig, database)([
-        ...new Set(scrapersInConfig)
-      ])
-      return () => {
-        const result = preparedStatment()
-        return groupUntilSeparator(
-          result,
-          ({ scraper }) => scraper === groupBy,
-          groupBy !== undefined && scrapers.includes(groupBy)
-        )
-      }
-    }
-    // create external query function
-    const query: QueryFn = params => prepare(params)()
-    // optionally call the `prepare` function first to get a prepared sqlite statement
-    query.prepare = prepare
-    return query
-  }
-}
-
-interface QueryFn {
-  (...args: ArgumentTypes<QueryFn['prepare']>): OrderedScrapersRow[][]
-  prepare: (
-    params: { scrapers: ScraperName[]; groupBy?: ScraperName }
-  ) => () => OrderedScrapersRow[][]
 }
 
 export { Store }
