+++ src/tools/store/querier-entrypoint/assert.runtime.ts
@@ -1,0 +1,7 @@
+import { assertType } from 'typescript-is'
+import { QueryArguments } from './'
+import { typescriptIsWrapper } from '../../../util/error'
+
+export const assertQueryArgumentsType = typescriptIsWrapper(queryArgs =>
+  assertType<QueryArguments>(queryArgs)
+)
