+++ src/util/rxjs/operators/index.ts
@@ -1,1 +1,2 @@
 export { rateLimitToggle } from './conditional-rate-limiter'
+export { wrapError } from './wrap-error'
